Renamed 'to-milliseconds' util file to ensure naming consistency
